% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaplessJDBC.R
\name{dbSettings}
\alias{dbSettings}
\title{Create an object to hold the database settings.}
\usage{
dbSettings(driverPath, driverName, address, user, password, javaBin = NULL)
}
\arguments{
\item{driverPath}{The local path to the JDBC driver.}

\item{driverName}{The name of the driver.}

\item{address}{The URL of the database.}

\item{user}{The user name to login to the database.}

\item{password}{The password to login to the database.}

\item{javaBin}{The directory where the java binary file is located in case this directory is not given in the OS environment variable called \code{'PATH'}.}
}
\value{
A \code{list} of class \code{"dbSettings"} which holds the settings.
}
\description{
Create an object to hold the database settings.
}
\examples{
# Example for an ORACLE database
dbSettings(
  driverPath = "C:/drivers/database/ojdbc6.jar",
  driverName = "oracle.jdbc.driver.OracleDriver",
  address = "jdbc:oracle:thin:@//yourInternal.Domain.com:1521/", # 1521 is the an example port
  user = "yourDbUserName",
  password = "yourDbUserPassword",
  javaBin = "C:/Program Files/Java/jre8/bin")
}
\author{
Daniel Hoop
}
