% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaplessJDBC.R
\name{queryToDataFrame}
\alias{queryToDataFrame}
\title{Execute a query and give back the result as a \code{data.frame}.}
\usage{
queryToDataFrame(statement, nullValue = "0", dbSettings, showCmd = FALSE)
}
\arguments{
\item{statement}{The SQL statement to execute.}

\item{nullValue}{The value which will be written into the file, if the database value was NULL (not zero!).}

\item{dbSettings}{An object of class \code{"dbSettings"} which holds the settings for the database. Must be created with the function \code{\link{dbSettings}}.}

\item{showCmd}{Logical value indicating if the command sent to an internal CLI should be shown. Only for debugging purposes.}
}
\value{
A \code{data.frame} which contains the result of the query.
}
\description{
Execute a query and give back the result as a \code{data.frame}.
}
\examples{
# Make the settings
# Example for an ORACLE database
dbs <- dbSettings(
  driverPath = "C:/drivers/database/ojdbc6.jar",
  driverName = "oracle.jdbc.driver.OracleDriver",
  address = "jdbc:oracle:thin:@//yourInternal.Domain.com:1521/", # 1521 is the an example port
  user = "yourDbUserName",
  password = "yourDbUserPassword",
  javaBin = "C:/Program Files/Java/jre8/bin")

# Execute
df <- queryToDataFrame(
  "SELECT * FROM persons WHERE age = 49",
  dbSettings = dbs)
}
\author{
Daniel Hoop
}
