% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaplessJDBC.R
\name{queryToFile}
\alias{queryToFile}
\title{Execute a query and write result to file.}
\usage{
queryToFile(
  statement,
  outputFile,
  sep = ";",
  nullValue = "0",
  dbSettings,
  checkOutputDir = TRUE,
  showCmd = FALSE
)
}
\arguments{
\item{statement}{The SQL statement to execute.}

\item{outputFile}{The file to write the result to.}

\item{sep}{The column separator in the file}

\item{nullValue}{The value which will be written into the file, if the database value was NULL (not zero!).}

\item{dbSettings}{An object of class \code{"dbSettings"} which holds the settings for the database. Must be created with the function \code{\link{dbSettings}}.}

\item{checkOutputDir}{Logical value indicating if the function should check whether the output directory exists before executing the command.}

\item{showCmd}{Logical value indicating if the command sent to an internal CLI should be shown. Only for debugging purposes.}
}
\value{
An error, if something went wrong, otherwise no return value.
}
\description{
Execute a query and write result to file.
}
\details{
See \code{\link{queryToDataFrame}} for an example.
}
\author{
Daniel Hoop
}
